pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dnsquest"
version = "1.0.0"
description = "An iterative DNS resolver built from scratch"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "dnspython>=2.4.0",
]

[project.optional-dependencies]
dev = [
    "black>=22.0.0",
    "pytest>=7.0.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
dnsquest = "dnsquest.__main__:main"

[tool.black]
line-length = 88
target-version = ['py37', 'py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | build
  | dist
  | __pycache__
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
